+++ spec/rubocop/cop/lint/unused_block_argument_spec.rb
@@ -172,20 +172,20 @@ describe RuboCop::Cop::Lint::UnusedBlockArgument do
       expect(autocorrect_source(cop, <<-SOURCE
       arr.map { |foo| stuff }
       SOURCE
-      )).to eq(<<-CORRECTED_SOURCE
+                               )).to eq(<<-CORRECTED_SOURCE
       arr.map { |_foo| stuff }
       CORRECTED_SOURCE
-      )
+                                       )
     end
 
     it 'fixes multiple' do
       expect(autocorrect_source(cop, <<-SOURCE
       hash.map { |key, val| stuff }
       SOURCE
-      )).to eq(<<-CORRECTED_SOURCE
+                               )).to eq(<<-CORRECTED_SOURCE
       hash.map { |_key, _val| stuff }
       CORRECTED_SOURCE
-      )
+                                       )
     end
 
     it 'preserves whitespace' do
@@ -193,31 +193,31 @@ describe RuboCop::Cop::Lint::UnusedBlockArgument do
       hash.map { |key,
                   val| stuff }
       SOURCE
-      )).to eq(<<-CORRECTED_SOURCE
+                               )).to eq(<<-CORRECTED_SOURCE
       hash.map { |_key,
                   _val| stuff }
       CORRECTED_SOURCE
-      )
+                                       )
     end
 
     it 'preserves splat' do
       expect(autocorrect_source(cop, <<-SOURCE
       obj.method { |foo, *bars, baz| stuff(foo, baz) }
       SOURCE
-      )).to eq(<<-CORRECTED_SOURCE
+                               )).to eq(<<-CORRECTED_SOURCE
       obj.method { |foo, *_bars, baz| stuff(foo, baz) }
       CORRECTED_SOURCE
-      )
+                                       )
     end
 
     it 'preserves default' do
       expect(autocorrect_source(cop, <<-SOURCE
       obj.method { |foo, bar = baz| stuff(foo) }
       SOURCE
-      )).to eq(<<-CORRECTED_SOURCE
+                               )).to eq(<<-CORRECTED_SOURCE
       obj.method { |foo, _bar = baz| stuff(foo) }
       CORRECTED_SOURCE
-      )
+                                       )
     end
 
     it 'ignores used' do